requirements.txt
markdown
django
namegenerator
names
django-tables2
flake8
django-mail-templated
unidecode
pyjwt